+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## Change Log
+**1 Maret 2022**
+- Remove default input value
+
 **20 Februari 2022**
 - Release